backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Werkzeug==3.0.1
WTForms==3.1.1
Pillow==10.1.0
python-dotenv==1.0.0
email-validator==2.1.0
Flask-CORS==4.0.1
flask-jwt-extended>=4.6.0
pydantic>=2.5.0,<3.0.0